+++ src/extension.ts
@@ -5,10 +5,11 @@ import * as geode from './geode/geode';
 import { browser } from './browser/browser';
 import { DevToolsPanel } from './devtools/DevToolsPanel';
 import { execSync } from 'child_process';
-import { getActiveProject } from './project/project';
+import { getActiveProject, getOpenedProjects } from './project/project';
 import { env } from 'vscode';
 import { Uri } from 'vscode';
 import { CCColorProvider } from './project/color';
+import { SpriteHoverPreview } from './project/sprite';
 
 export async function activate(context: ExtensionContext) {
 	const channel = window.createOutputChannel('Geode');
@@ -96,6 +97,9 @@ export async function activate(context: ExtensionContext) {
 	}));
 
 	context.subscriptions.push(languages.registerColorProvider({ language: "cpp" }, new CCColorProvider()));
+	context.subscriptions.push(languages.registerHoverProvider({ language: "cpp" }, new SpriteHoverPreview()));
+
+	getOutputChannel().appendLine(`Open Geode projects: ${getOpenedProjects().map(p => p.modJson.id).join(", ")}`);
 
 	// context.subscriptions.push(commands.registerCommand('geode.openDevTools', async () => {
 	// 	DevToolsPanel.show();